Move OpenXR input code into its own file (#858)
