+++ Dependencies/xr/CMakeLists.txt
@@ -33,7 +33,9 @@ else()
         "Source/OpenXR/XrRegistry.h"
         "Source/OpenXR/XrSupportedExtensions.h"
         "Source/OpenXR/SceneUnderstanding.h"
-        "Source/OpenXR/SceneUnderstanding.cpp")
+        "Source/OpenXR/SceneUnderstanding.cpp"
+        "Source/OpenXR/XrInput.h"
+        "Source/OpenXR/XrInput.cpp")
 endif()
 
 add_library(xr ${SOURCES})
